requirements.txt
sockets==1.0.0
cryptodome==3.10.1